upload now sends a message if a folder is found
